modified the hooks
